Move FunctionConfiguration out of Internal ns
Fix from review.
