+++ Solutions/Corvus.Testing.AzureFunctions/Corvus/Testing/AzureFunctions/FunctionConfiguration.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Endjin Limited. All rights reserved.
 // </copyright>
 
-namespace Corvus.Testing.AzureFunctions.Internal
+namespace Corvus.Testing.AzureFunctions
 {
     using System.Collections.Generic;
